Workaround activatability of Ayatana switch menu item
GtkMenuTrackerItem has the assumption that the GAction's parameter-type
and GMenuItem's "target" (i.e. "the parameter to pass when activating
the action" [1]) to match in type. This is reasonable, but Ayatana
Indicators' switch item doesn't follow it. Instead, it presents a
GMenuItem without a target, and a GAction with a boolean parameter.
ayatana-ido is then expected to pass the displaying switch's state as a
GAction's parameter.

Now, I'm not sure if breaking this assumption is the right thing to do,
but since the changes has shipped, we have to workaround it. This commit
makes a special case only for Ayatana Indicator's switch, so that it
won't break for anything else.

Workarounds https://github.com/AyatanaIndicators/qmenumodel/issues/21

[1] https://wiki.gnome.org/Projects/GLib/GApplication/DBusAPI#Attributes